Fix pull request comments
