+++ src/test/java/ru/cs/eatright/parsing/TokenizerTest.java
@@ -1,6 +1,10 @@
 package ru.cs.eatright.parsing;
 
 import org.junit.Test;
+import ru.cs.eatright.nlpmodel.StopWordTokenizer;
+import ru.cs.eatright.nlpmodel.Token;
+import ru.cs.eatright.nlpmodel.Tokenizer;
+
 import java.util.Arrays;
 import java.util.List;
